refactor: create `use-autocomplete.ts` for refactoring
