+++ src/components/search/search-panel.tsx
@@ -1,10 +1,9 @@
 "use client";
 
 import React from "react";
-import { createAutocomplete } from "@algolia/autocomplete-core";
-import { getAlgoliaResults } from "@algolia/autocomplete-preset-algolia";
+import { useRouter } from "next/navigation";
+import { File, Search } from "lucide-react";
 
-import { client as searchClient } from "@/lib/algolia";
 import {
   InternalSearchHit,
   InternalSearchHitWithParent,
@@ -12,9 +11,8 @@ import {
   InternalStoredSearchHit,
 } from "@/types/algolia";
 import { cn } from "@/lib/utils";
-import { useRouter } from "next/navigation";
-import { File, Search } from "lucide-react";
 import { createStoredSearches } from "./create-stored-searches";
+import useMemoizedAutocomplete from "./use-autocomplete";
 
 type Props = {
   isOpen: boolean;
@@ -69,100 +67,14 @@ function SearchPanel({ onClose }: Props): React.JSX.Element {
     [favoriteSearches, recentSearches]
   );
 
-  const autocomplete = React.useMemo(
-    () =>
-      createAutocomplete<
-        InternalSearchHitWithParent,
-        React.FormEvent<HTMLFormElement>,
-        React.MouseEvent,
-        React.KeyboardEvent
-      >({
-        id: "algolia-autocomplete",
-        placeholder: "Search articles...",
-        openOnFocus: true,
-        defaultActiveItemId: 0,
-        initialState: {
-          query: "",
-        },
-        onStateChange({ state }) {
-          // (2) Synchronize the Autocomplete state with the React state.
-          setAutocompleteState(state);
-        },
-        navigator: {
-          navigate({ itemUrl }) {
-            push(itemUrl);
-          },
-        },
-        getSources({ query }) {
-          if (!query) {
-            return [
-              {
-                sourceId: "recent searches",
-                getItemUrl({ item }) {
-                  return `/blog/${item.objectID}`;
-                },
-                getItems() {
-                  return recentSearches.getAll() as InternalSearchHitWithParent[];
-                },
-                onSelect({ item, event }) {
-                  saveRecentSearch(item);
-
-                  if (!isModifierEvent(event)) {
-                    onClose();
-                  }
-                },
-              },
-            ];
-          }
-
-          return [
-            // (3) Use an Algolia index source.
-            {
-              sourceId: "article",
-              getItemInputValue({ item }) {
-                return item.query as string;
-              },
-              onSelect({ item, event }) {
-                saveRecentSearch(item);
-
-                if (!isModifierEvent(event)) {
-                  onClose();
-                }
-              },
-              getItems({ query }) {
-                return getAlgoliaResults({
-                  searchClient,
-                  queries: [
-                    {
-                      indexName: indexName,
-                      params: {
-                        query,
-                        hitsPerPage: 4,
-                        highlightPreTag: "<mark>",
-                        highlightPostTag: "</mark>",
-                        snippetEllipsisText: "...",
-                        attributesToRetrieve: [
-                          "title",
-                          "imageUrl",
-                          "description",
-                          "tags",
-                          "headings",
-                          "publishedAt",
-                        ],
-                        attributesToSnippet: ["description:20", "headings:20"],
-                      },
-                    },
-                  ],
-                });
-              },
-              getItemUrl({ item }) {
-                return `/blog/${item.objectID}`;
-              },
-            },
-          ];
-        },
-      }),
-    [indexName, onClose, push, recentSearches, saveRecentSearch]
+  const autocomplete = useMemoizedAutocomplete(
+    autocompleteState,
+    setAutocompleteState,
+    indexName,
+    onClose,
+    recentSearches,
+    saveRecentSearch,
+    push
   );
 
   return (
@@ -220,7 +132,11 @@ function SearchPanel({ onClose }: Props): React.JSX.Element {
                             className="ais-source-item"
                             {...itemProps}
                           >
-                            {source.sourceId === "article" ? <File /> : <Search />}
+                            {source.sourceId === "article" ? (
+                              <File />
+                            ) : (
+                              <Search />
+                            )}
                             <span>{item.title}</span>
                           </li>
                         );
@@ -237,21 +153,3 @@ function SearchPanel({ onClose }: Props): React.JSX.Element {
 }
 
 export default SearchPanel;
-
-/**
- * Detect when an event is modified with a special key to let the browser
- * trigger its default behavior.
- */
-export function isModifierEvent<TEvent extends KeyboardEvent | MouseEvent>(
-  event: TEvent
-): boolean {
-  const isMiddleClick = (event as MouseEvent).button === 1;
-
-  return (
-    isMiddleClick ||
-    event.altKey ||
-    event.ctrlKey ||
-    event.metaKey ||
-    event.shiftKey
-  );
-}
